backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
langchain
langchain-openai
langchain-community
openai
httpx
pydantic
pydantic-settings
python-dotenv
cryptography
